Patch Note - oopsie, i accidentally deleted some of the patches that have already been made in last patch - Moved AdditionOfPowers in other java
